Fixed traversal of leaves in collection nodes.
Collections were being considered leaves in the tree, instead of
it's children. This led to unexpected behavior during validation
and hydration of data. This commit moves the responsibility of
walking the tree to each node, allowing polymorphism to take
place when needed.

I've also made a small refactoring on the way we handle required
fields. Now it's easier to customize that specific case.